Extract Agent SQL database budget
